+++ torrent-search-client/src/torrent.rs
@@ -4,7 +4,6 @@ use crate::{
 };
 use chrono::{DateTime, TimeZone, Utc};
 use derive_getters::Getters;
-use juniper::GraphQLObject;
 use serde::{Serialize, Serializer};
 use urlencoding::encode;
 
@@ -15,7 +14,7 @@ where
     serializer.serialize_str(&datetime.to_rfc3339())
 }
 
-#[derive(Serialize, Debug, Getters, Clone, GraphQLObject)]
+#[derive(Serialize, Debug, Getters, Clone)]
 pub struct Torrent {
     #[serde(serialize_with = "serialize_datetime")]
     pub added: DateTime<Utc>,
@@ -26,7 +25,7 @@ pub struct Torrent {
     pub leechers: i32,
     pub name: String,
     pub seeders: i32,
-    pub size: i32,
+    pub size: u64,
     pub provider: Provider,
     pub magnet: String,
     pub movie_properties: Option<MovieProperties>,
